bug fix tpm counter monitoring
bug fix tpm counter monitoring
